fix(dl-duplicate): minor updates
