--- libs/application/templates/driving-license-duplicate/src/fields/QualityPhoto/hooks/queries.graphql.ts
@@ -1,9 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const QUALITY_PHOTO = gql`
-  query HasQualityPhoto {
-    drivingLicenseQualityPhoto {
-      dataUri
-    }
-  }
-`
