--- libs/application/templates/driving-license-duplicate/src/fields/QualitySignature/hooks/queries.graphql.ts
@@ -1,9 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const QUALITY_SIGNATURE = gql`
-  query HasQualitySignature {
-    drivingLicenseQualitySignature {
-      dataUri
-    }
-  }
-`
