--- libs/application/templates/driving-license-duplicate/src/fields/SubmitAndDecline/index.ts
@@ -1,1 +1,0 @@
-export { SubmitAndDecline } from './SubmitAndDecline'
