+++ libs/application/templates/driving-license-duplicate/src/fields/index.ts
@@ -1,4 +1,4 @@
-export { QualityPhoto } from './QualityPhoto/index'
+export { QualityPhoto } from './QualityPhoto'
 export { QualitySignature } from './QualitySignature'
 export { CurrentLicense } from './CurrentLicense'
 export { PaymentCharge } from './PaymentCharge'
